--- src/main/java/com/kuit/findyou/domain/information/service/animalShelter/SyncVolunteerWorkService.java
@@ -1,8 +1,0 @@
-package com.kuit.findyou.domain.information.service.animalShelter;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface SyncVolunteerWorkService {
-    void synchronize();
-}
