--- src/main/java/com/kuit/findyou/domain/information/service/animalShelter/VolunteerWorkService.java
@@ -1,7 +1,0 @@
-package com.kuit.findyou.domain.information.service.animalShelter;
-
-import com.kuit.findyou.domain.information.dto.GetVolunteerWorksResponse;
-
-public interface VolunteerWorkService {
-    GetVolunteerWorksResponse getVolunteerWorksByCursor(Long lastId, int size);
-}
