+++ src/test/java/com/kuit/findyou/domain/information/service/VolunteerWorkServiceTest.java
@@ -3,7 +3,7 @@ package com.kuit.findyou.domain.information.service;
 import com.kuit.findyou.domain.information.dto.GetVolunteerWorksResponse;
 import com.kuit.findyou.domain.information.model.VolunteerWork;
 import com.kuit.findyou.domain.information.repository.VolunteerWorkRepository;
-import com.kuit.findyou.domain.information.service.animalShelter.VolunteerWorkServiceImpl;
+import com.kuit.findyou.domain.information.service.volunteerWork.VolunteerWorkServiceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
